add tests for vesting
