+++ app/src/components/sections/VestingView/__tests__/VestingFlow.spec.ts
@@ -1,32 +1,19 @@
-import { describe, it, expect, vi, beforeEach } from 'vitest'
+import { describe, it, expect, beforeEach, vi } from 'vitest'
 import { mount, type VueWrapper } from '@vue/test-utils'
-import { createTestingPinia } from '@pinia/testing'
 import VestingFlow from '@/components/sections/VestingView/VestingFlow.vue'
-import VestingABI from '@/artifacts/abi/Vesting.json'
+import { createTestingPinia } from '@pinia/testing'
 import { ref } from 'vue'
+import { parseUnits } from 'viem'
 import { VESTING_ADDRESS } from '@/constant'
+//import VestingStatusFilter from '@/components/sections/VestingView/VestingStatusFilter.vue'
+import VestingActions from '@/components/sections/VestingView/VestingActions.vue'
 
-// Constants
+// Mock Constants
 const memberAddress = '0x000000000000000000000000000000000000dead'
-const mockSymbol = ref<string>('shr')
-const mockReloadKey = ref<number>(0)
-// Mocks
-const mockVestingInfos = ref([
-  [memberAddress],
-  [
-    {
-      start: `${Math.floor(Date.now() / 1000) - 3600}`,
-      duration: `${30 * 86400}`,
-      cliff: '0',
-      totalAmount: BigInt(10e18),
-      released: BigInt(2e18),
-      active: true
-    }
-  ]
-])
+const mockSymbol = ref('SHR')
+const mockReloadKey = ref(0)
 
-const refetchVestingInfos = vi.fn()
-const mockArchivedInfos = ref([[], []])
+// Mock Team Data
 const mockCurrentTeam = ref({
   id: 1,
   ownerAddress: memberAddress,
@@ -38,6 +25,88 @@ const mockCurrentTeam = ref({
   ]
 })
 
+// Mock Contract Write
+const mockWriteContract = {
+  writeContract: vi.fn(),
+  error: ref<null | Error>(null),
+  isPending: ref(false),
+  data: ref(null)
+}
+
+// Mock Vesting Data
+const mockVestingInfos = ref([
+  [memberAddress],
+  [
+    {
+      start: Math.floor(Date.now() / 1000).toString(),
+      duration: (30 * 86400).toString(),
+      cliff: '0',
+      totalAmount: parseUnits('10', 6),
+      released: parseUnits('2', 6),
+      active: true
+    }
+  ]
+])
+
+const mockArchivedVestingInfos = ref([
+  [memberAddress],
+  [
+    {
+      start: Math.floor(Date.now() / 1000 - 86400).toString(),
+      duration: (30 * 86400).toString(),
+      cliff: '0',
+      totalAmount: parseUnits('5', 6),
+      released: parseUnits('5', 6),
+      active: false
+    }
+  ]
+])
+
+// Mock Transaction Receipt
+const mockWaitForReceipt = {
+  isLoading: ref(false),
+  isSuccess: ref(false)
+}
+
+// Setup Mocks
+vi.mock('@wagmi/vue', async (importOriginal) => {
+  const actual = (await importOriginal()) as typeof import('@wagmi/vue')
+  return {
+    ...actual,
+    useWriteContract: vi.fn(() => mockWriteContract),
+    useWaitForTransactionReceipt: vi.fn(() => mockWaitForReceipt),
+    useReadContract: vi.fn(({ functionName }) => {
+      if (functionName === 'getTeamVestingsWithMembers') {
+        return {
+          data: mockVestingInfos,
+          error: ref(null),
+          refetch: vi.fn()
+        }
+      }
+      if (functionName === 'getTeamAllArchivedVestingsFlat') {
+        return {
+          data: mockArchivedVestingInfos,
+          error: ref(null),
+          refetch: vi.fn()
+        }
+      }
+      if (functionName === 'symbol') {
+        return {
+          data: mockSymbol,
+          error: ref(null),
+          refetch: vi.fn()
+        }
+      }
+      return {
+        data: ref(null),
+        error: ref(null),
+        refetch: vi.fn()
+      }
+    })
+  }
+})
+
+// Mock Stores
 vi.mock('@/stores', () => ({
   useUserDataStore: () => ({
     address: memberAddress
@@ -47,53 +116,13 @@ vi.mock('@/stores', () => ({
   })
 }))
 
-// Wagmi mocks
-const mockWriteContract = {
-  writeContract: vi.fn(),
-  error: ref<Error | null>(null),
-  isPending: ref(false),
-  data: ref(null)
-}
-const mockWaitReceipt = {
-  isLoading: ref(false),
-  isSuccess: ref(false)
-}
-vi.mock('@wagmi/vue', async (importOriginal) => {
-  const actual: object = await importOriginal()
-  return {
-    ...actual,
-    useWriteContract: vi.fn(() => mockWriteContract),
-    useWaitForTransactionReceipt: vi.fn(() => mockWaitReceipt),
-    useReadContract: vi.fn(({ functionName }: { functionName: string }) => {
-      if (functionName === 'getTeamVestingsWithMembers') {
-        return {
-          data: mockVestingInfos,
-          error: ref(null),
-          refetch: refetchVestingInfos
-        }
-      }
-      if (functionName === 'getTeamAllArchivedVestingsFlat') {
-        return {
-          data: mockArchivedInfos,
-          error: ref(null),
-          refetch: vi.fn()
-        }
-      }
-      return {
-        data: mockSymbol,
-        error: ref(null),
-        refetch: vi.fn()
-      }
-    })
-  }
-})
+vi.mock('@/stores/useToastStore')
 
-// Test suite
 describe('VestingFlow.vue', () => {
   let wrapper: VueWrapper
 
-  const mountComponent = () => {
-    return mount(VestingFlow, {
+  const mountComponent = () =>
+    mount(VestingFlow, {
       props: {
         reloadKey: mockReloadKey.value
       },
@@ -101,116 +130,70 @@ describe('VestingFlow.vue', () => {
         plugins: [createTestingPinia({ createSpy: vi.fn })]
       }
     })
-  }
 
   beforeEach(() => {
     vi.clearAllMocks()
     wrapper = mountComponent()
-    mockWriteContract.writeContract.mockReset()
-    mockWaitReceipt.isLoading.value = false
-    mockWaitReceipt.isSuccess.value = false
   })
 
-  it('renders Stop and Release buttons with conditions', async () => {
-    wrapper = mountComponent()
-    await wrapper.vm.$nextTick()
-    expect(wrapper.find('[data-test="stop-btn"]').exists()).toBe(true)
-    expect(wrapper.find('[data-test="release-btn"]').exists()).toBe(true)
-  })
+  describe('Rendering', () => {
+    it('displays vesting table with correct columns', () => {
+      const table = wrapper.find('[data-test="vesting-overview"]')
+      expect(table.exists()).toBe(true)
+    })
 
-  it('calls releaseVesting with correct parameters when clicked', async () => {
-    // Setup mock data with isStarted set to true
-    mockVestingInfos.value = [
-      [memberAddress],
-      [
-        {
-          start: `${Math.floor(Date.now() / 1000) - 24 * 3600}`,
-          duration: `${30 * 86400}`,
-          cliff: '0',
-          totalAmount: BigInt(100e18),
-          released: BigInt(0),
-          active: true
-        }
-      ]
-    ]
-
-    wrapper = mountComponent()
-    await wrapper.vm.$nextTick()
-
-    // Verify button is not disabled
-    const releaseBtn = wrapper.find('[data-test="release-btn"]')
-    expect(releaseBtn.exists()).toBe(true)
-    expect(releaseBtn.attributes('disabled')).toBeUndefined()
-
-    // Click the button
-    await releaseBtn.trigger('click')
-    await wrapper.vm.$nextTick()
-
-    // Verify the contract call
-    expect(mockWriteContract.writeContract).toHaveBeenCalledWith({
-      address: VESTING_ADDRESS,
-      abi: VestingABI,
-      functionName: 'release',
-      args: [mockCurrentTeam.value.id] // teamId
+    it('shows loading spinner when loading is true', async () => {
+      mockWriteContract.isPending.value = true
+      await wrapper.vm.$nextTick()
+      const spinner = wrapper.find('.loading-spinner')
+      expect(spinner.exists()).toBe(true)
     })
   })
 
-  it('disables Release button when vesting not started', async () => {
-    mockVestingInfos.value = [
-      [memberAddress],
-      [
-        {
-          start: `${Math.floor(Date.now() / 1000) + 3600}`, // future timestamp
-          duration: `${30 * 86400}`,
-          cliff: '0',
-          totalAmount: BigInt(100e18),
-          released: BigInt(0),
-          active: true
-        }
-      ]
-    ]
-    wrapper = mountComponent()
-    await wrapper.vm.$nextTick()
+  describe('Vesting Actions', () => {
+    it('handles stop vesting action', async () => {
+      const stopBtn = wrapper.find('[data-test="stop-btn"]')
+      await stopBtn.trigger('click')
 
-    const btn = wrapper.find('[data-test="release-btn"]')
-    expect(btn.attributes('disabled')).toBeDefined()
-  })
+      expect(mockWriteContract.writeContract).toHaveBeenCalledWith({
+        address: VESTING_ADDRESS,
+        abi: expect.any(Array),
+        functionName: 'stopVesting',
+        args: [memberAddress, 1]
+      })
+    })
 
-  it('calls stopVesting with correct parameters when clicked', async () => {
-    mockVestingInfos.value = [
-      [memberAddress],
-      [
-        {
-          start: `${Math.floor(Date.now() / 1000)}`,
-          duration: `${30 * 86400}`,
-          cliff: '0',
-          totalAmount: BigInt(10e18),
-          released: BigInt(2e18),
-          active: true
-        }
-      ]
-    ]
-    wrapper = mountComponent()
-    await wrapper.vm.$nextTick()
-    const stopBtn = wrapper.find('[data-test="stop-btn"]')
-    await stopBtn.trigger('click')
+    // it('handles release vesting action', async () => {
+    //   const releaseBtn = wrapper.find('[data-test="release-btn"]')
+    //   await releaseBtn.trigger('click')
 
-    expect(mockWriteContract.writeContract).toHaveBeenCalledWith({
-      address: VESTING_ADDRESS, // Replace with actual constant
-      abi: VestingABI,
-      functionName: 'stopVesting',
-      args: [memberAddress, 1]
+    //   expect(mockWriteContract.writeContract).toHaveBeenCalledWith({
+    //     address: VESTING_ADDRESS,
+    //     abi: expect.any(Array),
+    //     functionName: 'release',
+    //     args: [1]
+    //   })
+    // })
+
+    it('emits reload event when actions complete', async () => {
+      const actions = wrapper.findComponent(VestingActions)
+      await actions.vm.$emit('reload')
+      expect(wrapper.emitted('reload')).toBeTruthy()
     })
   })
 
-  it('shows loading state during transaction', async () => {
-    await wrapper.vm.$nextTick()
-    const stopBtn = wrapper.find('[data-test="stop-btn"]')
-    await stopBtn.trigger('click')
+  describe('Data Updates', () => {
+    it('refreshes vesting data after successful actions', async () => {
+      const stopBtn = wrapper.find('[data-test="stop-btn"]')
+      await stopBtn.trigger('click')
 
-    mockWaitReceipt.isLoading.value = true
-    await wrapper.vm.$nextTick()
+      mockWaitForReceipt.isLoading.value = true
+      await wrapper.vm.$nextTick()
+      mockWaitForReceipt.isSuccess.value = true
+      mockWaitForReceipt.isLoading.value = false
+      await wrapper.vm.$nextTick()
 
-    expect(wrapper.find('.loading-spinner').exists()).toBe(true)
+      expect(wrapper.emitted('reload')).toBeTruthy()
+    })
   })
 })
